(internal) move example files, move src to js
