+++ example/bus/message.ts
