+++ example/index.ts
@@ -1,9 +1,10 @@
 import React from 'react';
 import { NewRoot } from './Root';
 import { createReconciler } from './moduleRenderer';
-import { createElement } from './manager/element';
 import { dispatch } from './bus/messageBus';
 import { TREE_MOUNT } from './bus/message';
+
+import { createElement } from '../src/manager/element';
 
 const renderer = createReconciler({ getInstance: () => null });
 const rootElement = createElement('root', {});
