+++ example/types/plugin.d.ts
