--- src/data/event.ts
@@ -1,6 +1,0 @@
-export type Event<P = unknown, M = unknown> = {
-    type: string
-    payload: P
-    meta?: M
-}
-
