+++ src/log-level.js
@@ -1,3 +1,6 @@
+/** @import {LogSeverity} from './types' */
+
+/** @type {Record<string, LogSeverity} */
 export const LOG_LEVEL = {
     FATAL: 1,
     ERROR: 2,
@@ -5,6 +8,4 @@ export const LOG_LEVEL = {
     INFO: 4,
     DEBUG: 5,
     TRACE: 6,
-} as const
-
-export type LogSeverity = typeof LOG_LEVEL[keyof typeof LOG_LEVEL]
+}
